fix(plugins): report supported bundle hook layouts
Share metadata discovery with the runtime hook loader so detected JSON automation is not advertised as a supported hook. Preserve declared capabilities and valid mixed bundles; remove the old unchecked manifest cast.
