+++ src/hooks/workspace.ts
@@ -1,42 +1,20 @@
 // Hook workspace helpers resolve hook roots and workspace-local hook files.
-import fs from "node:fs";
 import path from "node:path";
-import { safeParseJson } from "@openclaw/normalization-core";
 import { normalizeTrimmedStringList } from "@openclaw/normalization-core/string-normalization";
-import { parseFrontmatterBlockResult } from "../../packages/markdown-core/src/frontmatter.js";
-import { MANIFEST_KEY } from "../compat/legacy-names.js";
 import type { OpenClawConfig } from "../config/types.openclaw.js";
-import { openRootFileSync, readFileDescriptorBoundedSync } from "../infra/boundary-file-read.js";
 import { createSubsystemLogger } from "../logging/subsystem.js";
-import { isPathInsideWithRealpath } from "../security/scan-paths.js";
 import { CONFIG_DIR, resolveUserPath } from "../utils.js";
 import { resolveBundledHooksDir } from "./bundled-dir.js";
-import { resolveHookInvocationPolicy, resolveHookManifestMetadata } from "./frontmatter.js";
+import {
+  loadHookEntriesFromDir,
+  type DiscoveredHookEntry,
+  type HookDiscoveryRoot,
+} from "./discovery.js";
 import { resolvePluginHookDirs } from "./plugin-hooks.js";
 import { resolveHookEntries } from "./policy.js";
-import type { Hook, HookEntry, HookPolicyEntry, HookSource } from "./types.js";
+import type { HookEntry, HookPolicyEntry } from "./types.js";
 
-// Hook descriptors are small metadata. Bounding the pinned descriptor read also
-// covers files that grow after the boundary open validates their identity.
-const HOOK_METADATA_MAX_BYTES = 1024 * 1024;
-
-type HookPackageManifest = {
-  name?: string;
-} & Partial<Record<typeof MANIFEST_KEY, { hooks?: string[] }>>;
 const log = createSubsystemLogger("hooks/workspace");
-
-type DiscoveredHookEntry = Omit<HookEntry, "hook"> & {
-  hook: Omit<Hook, "handlerPath"> & { handlerPath?: string };
-  invalidMetadata?: boolean;
-};
-
-type HookDiscoveryRoot = {
-  dir: string;
-  source: HookSource;
-  pluginId?: string;
-  rootDir?: string;
-  includeRoot?: boolean;
-};
 
 export type HookSourceFact = HookPolicyEntry & { rootId: string; filePath: string };
 type HookCandidate = HookSourceFact & { entry?: DiscoveredHookEntry };
@@ -45,174 +23,6 @@ type HookDiscoveryOptions = {
   managedHooksDir?: string;
   bundledHooksDir?: string;
 };
-
-function readHookPackageManifest(dir: string): HookPackageManifest | null {
-  const manifestPath = path.join(dir, "package.json");
-  const raw = readRootFileUtf8({
-    absolutePath: manifestPath,
-    rootPath: dir,
-    boundaryLabel: "hook package directory",
-    maxBytes: HOOK_METADATA_MAX_BYTES,
-  });
-  if (raw === null) {
-    return null;
-  }
-  return (safeParseJson(raw) as HookPackageManifest | undefined) ?? null;
-}
-
-function resolvePackageHooks(manifest: HookPackageManifest): string[] {
-  return normalizeTrimmedStringList(manifest[MANIFEST_KEY]?.hooks);
-}
-
-function resolveContainedDir(baseDir: string, targetDir: string): string | null {
-  const base = path.resolve(baseDir);
-  const resolved = path.resolve(baseDir, targetDir);
-  if (
-    !isPathInsideWithRealpath(base, resolved, {
-      requireRealpath: true,
-    })
-  ) {
-    return null;
-  }
-  return resolved;
-}
-
-function loadHookFromDir(params: {
-  hookDir: string;
-  source: HookSource;
-  pluginId?: string;
-}): DiscoveredHookEntry | null {
-  const hookMdPath = path.join(params.hookDir, "HOOK.md");
-  const content = readRootFileUtf8({
-    absolutePath: hookMdPath,
-    rootPath: params.hookDir,
-    boundaryLabel: "hook directory",
-    maxBytes: HOOK_METADATA_MAX_BYTES,
-  });
-  if (content === null) {
-    return null;
-  }
-  try {
-    const { frontmatter, issues } = parseFrontmatterBlockResult(content);
-
-    const name = frontmatter.name || path.basename(params.hookDir);
-    const description = frontmatter.description || "";
-
-    const handlerCandidates = ["handler.ts", "handler.js", "index.ts", "index.js"];
-    let handlerPath: string | undefined;
-    for (const candidate of handlerCandidates) {
-      const candidatePath = path.join(params.hookDir, candidate);
-      const safeCandidatePath = resolveRootFilePath({
-        absolutePath: candidatePath,
-        rootPath: params.hookDir,
-        boundaryLabel: "hook directory",
-      });
-      if (safeCandidatePath) {
-        handlerPath = safeCandidatePath;
-        break;
-      }
-    }
-
-    if (!handlerPath) {
-      log.warn(`Hook "${name}" has HOOK.md but no readable handler in ${params.hookDir}`);
-    }
-
-    let baseDir = params.hookDir;
-    try {
-      baseDir = fs.realpathSync.native(params.hookDir);
-    } catch {
-      // keep the discovered path when realpath is unavailable
-    }
-
-    return {
-      hook: {
-        name,
-        description,
-        source: params.source,
-        pluginId: params.pluginId,
-        filePath: hookMdPath,
-        baseDir,
-        handlerPath,
-      },
-      frontmatter,
-      invalidMetadata: issues.length > 0,
-      metadata: resolveHookManifestMetadata(frontmatter),
-      invocation: resolveHookInvocationPolicy(frontmatter),
-    };
-  } catch (err) {
-    const message = err instanceof Error ? (err.stack ?? err.message) : String(err);
-    log.warn(`Failed to load hook from ${params.hookDir}: ${message}`);
-    return null;
-  }
-}
-
-function loadHooksFromCandidate(params: {
-  hookDir: string;
-  source: HookSource;
-  pluginId?: string;
-}): DiscoveredHookEntry[] | null {
-  const { hookDir, source, pluginId } = params;
-  const manifest = readHookPackageManifest(hookDir);
-  const packageHooks = manifest ? resolvePackageHooks(manifest) : [];
-  if (packageHooks.length === 0) {
-    if (!fs.existsSync(path.join(hookDir, "HOOK.md"))) {
-      return null;
-    }
-    const hook = loadHookFromDir(params);
-    return hook ? [hook] : [];
-  }
-
-  const hooks: DiscoveredHookEntry[] = [];
-  for (const hookPath of packageHooks) {
-    const resolvedHookDir = resolveContainedDir(hookDir, hookPath);
-    if (!resolvedHookDir) {
-      log.warn(
-        `Ignoring out-of-package hook path "${hookPath}" in ${hookDir} (must be within package directory)`,
-      );
-      continue;
-    }
-    // Pack entries are hook leaves, never another pack or a collection to scan.
-    const hook = loadHookFromDir({
-      hookDir: resolvedHookDir,
-      source,
-      pluginId,
-    });
-    if (hook) {
-      hooks.push(hook);
-    }
-  }
-
-  return hooks;
-}
-
-function loadHookEntriesFromDir(params: HookDiscoveryRoot): DiscoveredHookEntry[] {
-  const { dir, source, pluginId } = params;
-  // Plugin policy selects roots even when their files disappear. Boundary checks
-  // belong to discovery, so atomic reload can retain the selected source fact.
-  if (params.rootDir && !isPathInsideWithRealpath(params.rootDir, dir, { requireRealpath: true })) {
-    log.warn(`Plugin hook path is missing or escapes plugin root (${pluginId}): ${dir}`);
-    return [];
-  }
-  if (!fs.existsSync(dir) || !fs.statSync(dir).isDirectory()) {
-    return [];
-  }
-  const rootHooks = params.includeRoot
-    ? loadHooksFromCandidate({ hookDir: dir, source, pluginId })
-    : null;
-  // null means a collection. A recognized root with rejected hooks stays empty;
-  // falling back to children would execute code its manifest did not select.
-  return (
-    rootHooks ??
-    fs.readdirSync(dir, { withFileTypes: true }).flatMap((entry) => {
-      if (!entry.isDirectory()) {
-        return [];
-      }
-      return (
-        loadHooksFromCandidate({ hookDir: path.join(dir, entry.name), source, pluginId }) ?? []
-      );
-    })
-  );
-}
 
 function resolveHookDiscoveryRoots(
   workspaceDir: string,
@@ -255,7 +65,7 @@ export function prepareWorkspaceHookEntries(
       Boolean(root.includeRoot),
       root.rootDir,
     ]);
-    const entries: HookCandidate[] = loadHookEntriesFromDir(root).map((entry) => ({
+    const entries: HookCandidate[] = loadHookEntriesFromDir(root, log.warn).map((entry) => ({
       rootId,
       filePath: entry.hook.filePath,
       hook: { name: entry.hook.name, source: entry.hook.source },
@@ -328,58 +138,4 @@ export function loadWorkspaceHookEntries(
   opts?: HookDiscoveryOptions,
 ): HookEntry[] {
   return prepareWorkspaceHookEntries(workspaceDir, opts).entries;
-}
-
-function readRootFileUtf8(params: {
-  absolutePath: string;
-  rootPath: string;
-  boundaryLabel: string;
-  maxBytes: number;
-}): string | null {
-  return withOpenedRootFileSync(params, (opened) => {
-    try {
-      return readFileDescriptorBoundedSync(opened.fd, params.maxBytes).toString("utf-8");
-    } catch (err) {
-      if (err instanceof RangeError) {
-        log.warn(
-          `Ignoring oversized hook metadata ${params.absolutePath}: file exceeds the ${params.maxBytes}-byte limit`,
-        );
-      }
-      return null;
-    }
-  });
-}
-
-function withOpenedRootFileSync<T>(
-  params: {
-    absolutePath: string;
-    rootPath: string;
-    boundaryLabel: string;
-  },
-  read: (opened: { fd: number; path: string }) => T,
-): T | null {
-  const opened = openRootFileSync({
-    absolutePath: params.absolutePath,
-    rootPath: params.rootPath,
-    boundaryLabel: params.boundaryLabel,
-    // Operator hook dirs are commonly symlinked; fs-safe still rejects hops
-    // whose canonical target escapes the hook root.
-    rejectSymlinks: false,
-  });
-  if (!opened.ok) {
-    return null;
-  }
-  try {
-    return read({ fd: opened.fd, path: opened.path });
-  } finally {
-    fs.closeSync(opened.fd);
-  }
-}
-
-function resolveRootFilePath(params: {
-  absolutePath: string;
-  rootPath: string;
-  boundaryLabel: string;
-}): string | null {
-  return withOpenedRootFileSync(params, (opened) => opened.path);
 }
