+++ src/plugins/installed-plugin-components.ts
@@ -1,7 +1,9 @@
+import path from "node:path";
 import type {
   PluginDeclaredSurface,
   PluginInstalledComponents,
 } from "../../packages/gateway-protocol/src/schema/plugins.js";
+import { loadHookEntriesFromDir } from "../hooks/discovery.js";
 import { inspectBundleLspRuntimeSupport } from "./bundle-lsp.js";
 import {
   inspectBundleMcpRuntimeSupport,
@@ -68,6 +70,23 @@ export function projectInstalledPluginComponents(params: {
     capabilities: manifest.bundleCapabilities ?? [],
   });
   const mapped = new Set(support.mapped);
+  const hooks =
+    mapped.has("hooks") && manifest.rootDir
+      ? sorted(
+          (manifest.hooks ?? []).filter((dir) =>
+            loadHookEntriesFromDir({
+              dir: path.resolve(manifest.rootDir, dir),
+              rootDir: manifest.rootDir,
+              pluginId: manifest.id,
+              source: "openclaw-plugin",
+            }).some(({ hook }) => Boolean(hook.handlerPath)),
+          ),
+        )
+      : [];
+  if (mapped.has("hooks") && hooks.length === 0) {
+    mapped.delete("hooks");
+    support.unavailable.push("hooks");
+  }
   const mcp = manifest.rootDir
     ? inspectBundleMcpRuntimeSupport({
         pluginId: manifest.id,
@@ -87,7 +106,7 @@ export function projectInstalledPluginComponents(params: {
     skills: mapped.has("skills") ? sorted(declared.skills) : [],
     mcpServers: mapped.has("mcpServers") ? sorted(mcp?.supportedServerNames ?? []) : [],
     commands: [],
-    hooks: mapped.has("hooks") ? sorted(declared.hooks) : [],
+    hooks,
     lspServers: mapped.has("lspServers") ? sorted(lsp?.supportedServerNames ?? []) : [],
     unavailable: {
       capabilities: sorted(support.unavailable),
